requirements.txt
tqdm
xxhash
python-magic-bin